% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_types.R
\name{apply_solution}
\alias{apply_solution}
\title{Apply a Solution}
\usage{
apply_solution(x, soln)
}
\arguments{
\item{x}{The constraint set or solution set.}

\item{soln}{The solution to apply.}
}
\description{
This function substitutes a solution into a constraint set or solution set.
}

