% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASTWalkerState.R
\name{ASTWalkerState}
\alias{ASTWalkerState}
\title{AST Walker State}
\usage{
ASTWalkerState()
}
\description{
This R6 class stores the state of the AST walker, in order to signal to the
walker when and how far to ascend from an AST node.
}

