% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrain.R
\name{constrain}
\alias{constrain}
\title{Generate Type Constraint Set from Expression}
\usage{
constrain(node, helper = InferHelper())
}
\arguments{
\item{node}{ASTNode. Expression from which to generate type constraints.}

\item{helper}{InferHelper. Helper object for constraint generation.}
}
\value{
A list with two elements. The first is the list of constraints, and
the second is the final state of the helper object.
}
\description{
This function generates a set of type constraints from the given expression,
which can then be solved to infer types for the expression.
}
