% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_types.R
\name{detect_types}
\alias{detect_types}
\title{Type Detection}
\usage{
detect_types(node, set = ConstraintSet$new())
}
\arguments{
\item{node}{(ASTNode) An abstract syntax tree.}
}
\description{
This function detects and marks the types of nodes in a tree of ASTNodes
in-place. The return value includes a list of constraint tuples meant to be
solved via unification as part of the type inference process.
}

