% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol_map.R
\name{get_parameters}
\alias{get_parameters}
\title{Get All Parameters in a SymbolMap}
\usage{
get_parameters(map, include_na = FALSE)
}
\arguments{
\item{map}{(SymbolMap) The map to search for parameter entries.}

\item{include_na}{(logical) Should entries with `is_parameter = NA` be
included in the result?}
}
\value{
A list of entries.
}
\description{
This function gets all entries in a `SymbolMap` that correspond to function
parameters.
}
