% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_dm.R
\name{infer_dm}
\alias{infer_dm}
\title{Infer Types for a Function}
\usage{
infer_dm(node, env, counter, param_types)
}
\arguments{
\item{node}{(Function) The function to infer types for.}

\item{counter}{(Counter) A counter for naming type variables.}

\item{tenv}{(TypeEnvironment) The global type environment.}
}
\description{
This function infers the types for a function.
}
