% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_types.R
\name{infer_types}
\alias{infer_types}
\title{Infer Types for a Function}
\usage{
infer_types(code, init = list(), scalar = FALSE,
  constraintHandlers = getConstraintHandlers(),
  solveHandlers = getSolveHandlers(), ...)
}
\description{
Infer Types for a Function
}
