% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_helper.R
\docType{data}
\name{InferHelper}
\alias{InferHelper}
\title{Infer Helper}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
InferHelper
}
\description{
Infer Helper
}
\keyword{datasets}
