% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferTypes.R
\name{inferTypes}
\alias{inferTypes}
\title{Infer Types}
\usage{
inferTypes(x, typeCollector = TypeCollector(), ...)
}
\arguments{
\item{x}{unquoted R expression}

\item{typeCollector}{TypeCollector object}

\item{...}{additional arguments}
}
\description{
This function infers the types of an expression.
}

