% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{return_type}
\alias{return_type}
\title{Get Return Type from Solution Set}
\usage{
return_type(solutions)
}
\arguments{
\item{(SolutionSet)}{Complete type information for a function.}
}
\description{
This function gets the return type from the solved types for a function.
}
