% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{solve.ConstraintSet}
\alias{solve.ConstraintSet}
\title{Solve a Set of Type Constraints}
\usage{
\method{solve}{ConstraintSet}(set)
}
\arguments{
\item{set}{(ConstraintSet) A set of type constraints.}
}
\description{
Solve a Set of Type Constraints
}

