% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TypeCollector.R
\docType{class}
\name{TypeCollector-class}
\alias{TypeCollector}
\alias{TypeCollector-class}
\title{TypeCollector}
\description{
This reference class stores type information gathered by the type inference
algorithm.
}

