% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator_handlers.R
\name{upcast_math}
\alias{upcast_math}
\title{Upcast Types for Math Operator}
\usage{
upcast_math(types, op)
}
\arguments{
\item{types}{(list) The input types.}

\item{op}{(character) The name of the math operator.}
}
\description{
This function determines the output type of a unary or binary math operator
when given the input types.
}
