% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrain.R
\name{constrain}
\alias{constrain}
\title{Generate Type Constraint Set from Expression}
\usage{
constrain(node, map = SymbolMap())
}
\arguments{
\item{node}{(ASTNode) An expression from which to generate type constraints.}

\item{map}{(SymbolMap) A mapping between program variables and type
variables.}
}
\value{
An S4 \code{Result} object, with slots \code{contraints} and \code{map}. The former
is the list of constraints, and the latter is the mapping between program
variables and type variables.
}
\description{
This function generates a set of type constraints from the given expression,
which can then be solved to infer types for the expression.
}
