% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrain.R
\name{constrain}
\alias{constrain}
\title{Compute Type Constraints from a Control-flow Graph}
\usage{
constrain(cfg, b, dom_t = rstatic::dom_tree(cfg), set = ConstraintSet$new(),
  scalar = FALSE, handlers = getConstraintHandlers())
}
\description{
Compute Type Constraints from a Control-flow Graph
}
