% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrain.R
\name{constrain}
\alias{constrain}
\title{Compute Type Constraints from a Control-flow Graph}
\usage{
constrain(cfg, b = 1L, dom_t = ast::dom_tree(cfg),
  set = ConstraintSet$new())
}
\description{
Compute Type Constraints from a Control-flow Graph
}
