% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_types.R
\name{infer_types}
\alias{infer_types}
\title{Infer Types}
\usage{
infer_types(expr, scope = Scope(), ...)
}
\arguments{
\item{expr}{unquoted R expression}

\item{scope}{Scope object}

\item{...}{additional arguments}
}
\description{
This function infers the types of an expression.
}

